feat(mcp): extract search ranking helpers into search-ranking-lib
Move MCP registry query normalization, platform matching, trust signals,
and ranked search scoring into a dedicated lib module with wrapper
re-exports and focused tests.
